src/2. BigData/3. Deployment/3-VM/backend/pyproject.toml
[project]
name = "iowa-timeseries-minimal"
version = "0.1.0"
description = "Minimal AutoGluon time series forecasting example"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0,<2.0.0",
    "torch>=2.2.0,<2.3.0",
    "autogluon.timeseries>=1.0.0",
    "matplotlib>=3.7.0",
    "google-cloud-bigquery>=3.0.0",
    "db-dtypes>=1.0.0",
]
